% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenokio.R
\name{.resolve_color_bgr}
\alias{.resolve_color_bgr}
\title{Resolve any R color name or BGR vector to OpenCV-compatible BGR}
\usage{
.resolve_color_bgr(col)
}
\description{
Resolve any R color name or BGR vector to OpenCV-compatible BGR
}
