% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenokio.R
\name{ensure_pydeps}
\alias{ensure_pydeps}
\title{Internal helper to ensure Python dependencies are installed}
\usage{
ensure_pydeps()
}
\description{
Internal helper to ensure Python dependencies are installed
}
